requirements.txt
certifi==2019.11.28
cycler==0.10.0
kiwisolver==1.1.0
matplotlib==3.2.0
numpy==1.18.1
Pillow==7.0.0
pyparsing==2.4.6
python-dateutil==2.8.1
six==1.14.0
